Correctly do browser cleanup
Kinda refactored the code :)
