+++ chatgpt_wrapper/chatgpt.py
@@ -1,4 +1,3 @@
-import atexit
 import base64
 import json
 from json.decoder import JSONDecodeError
@@ -14,6 +13,8 @@ from typing import Optional
 from playwright.sync_api import sync_playwright
 from playwright._impl._api_structures import ProxySettings
 from playwright.sync_api._generated import BrowserContext
+from .log import LogCapability
+from .browser import Browser
 
 RENDER_MODELS = {
     "default": "text-davinci-002-render-sha",
@@ -26,7 +27,7 @@ DEFAULT_CONSOLE_LOG_FORMATTER = logging.Formatter("%(levelname)s - %(message)s")
 DEFAULT_FILE_LOG_LEVEL = logging.DEBUG
 DEFAULT_FILE_LOG_FORMATTER = logging.Formatter("%(asctime)s - %(levelname)s - %(message)s")
 
-class ChatGPT:
+class ChatGPT(LogCapability):
     """
     A ChatGPT interface that uses Playwright to run a browser,
     and interacts with that browser to communicate with ChatGPT in
@@ -36,139 +37,27 @@ class ChatGPT:
     stream_div_id = "chatgpt-wrapper-conversation-stream-data"
     eof_div_id = "chatgpt-wrapper-conversation-stream-data-eof"
     session_div_id = "chatgpt-wrapper-session-data"
-    session={}
-    # A single play can launch many browsers
-    play=sync_playwright().start()
+    # The default browser
+    default_browser=Browser()
 
-    def __init__(self, headless: bool = True, browser="firefox", model="default", timeout=60, debug_log=None, proxy: Optional[ProxySettings] = None):
-        self.log = self._set_logging(debug_log)
+    def __init__(self, headless: bool = True, browser=default_browser, model="default", timeout=60, debug_log=None, proxy: Optional[ProxySettings] = None):
+        super().__init__(debug_log=debug_log)
         self.log.info("ChatGPT initialized")
 
-        self.browser=self.page=None # Without this the linter would complain about defining attributes outside init
-        self._setup_browser(headless=headless,browser=browser,proxy=proxy)
-
-        self._start_browser()
+        self.browser=browser
         self.parent_message_id = str(uuid.uuid4())
         self.conversation_id = None
         self.conversation_title_set = None
         self.model = model
         self.timeout = timeout
-        atexit.register(self._cleanup)
-
-    def _setup_browser(self,headless: bool, browser, proxy: ProxySettings):
-        """Setup the browser
-
-        Args:
-            headless (bool)
-            browser (str | BrowserContext):
-                If it is str, it should be one of 'firefox', 'chromium' and 'webkit'.
-                If it is a BrowserContext(playwright.sync_api._generated.BrowserContext), it will be directly used as the browser and no new browser context is created.
-            proxy (ProxySettings): _description_
-        """
-        if isinstance(browser,str):
-            if not hasattr(self.play,browser):
-                self.log.error("Browser %s is invalid, falling back on firefox. ",browser)
-                browser='firefox'
-            playbrowser=getattr(self.play,browser)
-
-            try:
-                self.browser=playbrowser.launch_persistent_context(
-                    user_data_dir="/tmp/playwright",
-                    headless=headless,
-                    proxy=proxy,
-                )
-            except Exception:
-                self.user_data_dir=f"/tmp/{str(uuid.uuid4())}"
-                shutil.copytree("/tmp/playwright", self.user_data_dir)
-                self.browser = playbrowser.launch_persistent_context(
-                    user_data_dir=self.user_data_dir,
-                    headless=headless,
-                    proxy=proxy,
-                )
-        elif isinstance(browser,BrowserContext):
-            self.browser=browser
-        else:
-            raise TypeError("Argument `browser` is neither str nor playwright.sync_api._generated.BrowserType")
-
-        if len(self.browser.pages)>0:
-            self.page=self.browser.pages[0]
-        else:
-            self.page=self.browser.new_page()
-        self._start_browser()
-
-    def _set_logging(self, debug_log):
-        logger = logging.getLogger(self.__class__.__name__)
-        logger.setLevel(logging.DEBUG)
-        log_console_handler = logging.StreamHandler()
-        log_console_handler.setFormatter(DEFAULT_CONSOLE_LOG_FORMATTER)
-        log_console_handler.setLevel(DEFAULT_CONSOLE_LOG_LEVEL)
-        logger.addHandler(log_console_handler)
-        if debug_log:
-            log_file_handler = logging.FileHandler(debug_log)
-            log_file_handler.setFormatter(DEFAULT_FILE_LOG_FORMATTER)
-            log_file_handler.setLevel(DEFAULT_FILE_LOG_LEVEL)
-            logger.addHandler(log_file_handler)
-        return logger
-
-    def _start_browser(self):
-        self.page.goto("https://chat.openai.com/")
-
-    def _cleanup(self):
-        # remove the user data dir in case this is a second instance
-        if hasattr(self, "user_data_dir"):
-            shutil.rmtree(self.user_data_dir)
-
-    def refresh_session(self,timeout=15):
-        """Refresh session, by redirecting the *page* to /api/auth/session rather than a simple xhr request.
-
-        In this way, we can pass the browser check.
-
-        Args:
-            timeout (int, optional): Timeout waiting for the refresh in seconds. Defaults to 10.
-        """
-        self.log.info("Refreshing session...")
-        self.page.goto("https://chat.openai.com/api/auth/session")
-        try:
-            self.page.wait_for_url("/api/auth/session",timeout=timeout*1000)
-        except Exception:
-            self.log.error("Timed out refreshing session. Page is now at %s. Calling _start_browser()...")
-            self._start_browser()
-        try:
-            while "Please stand by, while we are checking your browser..." in self.page.content():
-                time.sleep(1)
-            contents=self.page.content()
-            """
-            By GETting /api/auth/session, the server would ultimately return a raw json file.
-            However, as this is a browser, it will add something to it, like <body> or so, like this:
-
-            <html><head><link rel="stylesheet" href="resource://content-accessible/plaintext.css"></head><body><pre>{xxx:"xxx",{},accessToken="sdjlsfdkjnsldkjfslawefkwnlsdw"}
-            </pre></body></html>
-
-        session_data = json.loads(session_datas[0].inner_text())
-        ChatGPT.session = session_data
-            The following code tries to extract the json part from the page, by simply finding the first `{` and the last `}`.
-            """
-            found_json=re.search('{.*}',contents)
-            if found_json==None:
-                raise JSONDecodeError("Cannot find JSON in /api/auth/session 's response")
-            contents=contents[found_json.start():found_json.end()]
-            self.log.debug("Refreshing session received: %s",contents)
-            self.session=json.loads(contents)
-            self.log.info("Succeessfully refreshed session. ")
-        except json.JSONDecodeError:
-            self.log.error("Failed to decode session key. Maybe Access denied? ")
-
-        # Now the browser should be at /api/auth/session
-        # Go back to the chat page.
-        self._start_browser()
 
     def _cleanup_divs(self):
-        self.page.evaluate(f"document.getElementById('{self.stream_div_id}').remove()")
-        self.page.evaluate(f"document.getElementById('{self.eof_div_id}').remove()")
+        self.browser.page.evaluate(f"document.getElementById('{self.stream_div_id}').remove()")
+        self.browser.page.evaluate(f"document.getElementById('{self.eof_div_id}').remove()")
 
     def _api_request_build_headers(self, custom_headers={}):
         headers = {
-            "Authorization": "Bearer %s" % ChatGPT.session["accessToken"],
+            "Authorization": "Bearer %s" % self.browser.session["accessToken"],
         }
         headers.update(custom_headers)
         return headers
@@ -187,17 +76,17 @@ class ChatGPT:
 
     def _api_get_request(self, url, query_params={}, custom_headers={}):
         headers = self._api_request_build_headers(custom_headers)
-        response = self.page.request.get(url, headers=headers, params=query_params)
+        response = self.browser.page.request.get(url, headers=headers, params=query_params)
         return self._process_api_response(url, response)
 
     def _api_post_request(self, url, data={}, custom_headers={}):
         headers = self._api_request_build_headers(custom_headers)
-        response = self.page.request.post(url, headers=headers, data=data)
+        response = self.browser.page.request.post(url, headers=headers, data=data)
         return self._process_api_response(url, response, method="POST")
 
     def _api_patch_request(self, url, data={}, custom_headers={}):
         headers = self._api_request_build_headers(custom_headers)
-        response = self.page.request.patch(url, headers=headers, data=data)
+        response = self.browser.page.request.patch(url, headers=headers, data=data)
         return self._process_api_response(url, response, method="PATCH")
 
     def _gen_title(self):
@@ -230,8 +119,8 @@ class ChatGPT:
             parent_id = current_item['id']
 
     def delete_conversation(self, uuid=None):
-        if 'accessToken' not in ChatGPT.session:
-            self.refresh_session()
+        if 'accessToken' not in self.browser.session:
+            self.browser.refresh_session()
         if not uuid and not self.conversation_id:
             return
         id = uuid if uuid else self.conversation_id
@@ -246,8 +135,8 @@ class ChatGPT:
             self.log.error("Failed to delete conversation")
 
     def set_title(self, title, conversation_id=None):
-        if 'accessToken' not in ChatGPT.session:
-            self.refresh_session()
+        if 'accessToken' not in self.browser.session:
+            self.browser.refresh_session()
         id = conversation_id if conversation_id else self.conversation_id
         url = f"https://chat.openai.com/backend-api/conversation/{id}"
         data = {
@@ -260,8 +149,8 @@ class ChatGPT:
             self.log.error("Failed to set title")
 
     def get_history(self, limit=20, offset=0):
-        if 'accessToken' not in ChatGPT.session:
-            self.refresh_session()
+        if 'accessToken' not in self.browser.session:
+            self.browser.refresh_session()
         url = "https://chat.openai.com/backend-api/conversations"
         query_params = {
             "offset": offset,
@@ -277,8 +166,8 @@ class ChatGPT:
             self.log.error("Failed to get history")
 
     def get_conversation(self, id=None):
-        if 'accessToken' not in ChatGPT.session:
-            self.refresh_session()
+        if 'accessToken' not in self.browser.session:
+            self.browser.refresh_session()
         id = id if id else self.conversation_id
         if id:
             url = f"https://chat.openai.com/backend-api/conversation/{id}"
@@ -289,12 +178,12 @@ class ChatGPT:
                 self.log.error(f"Failed to get conversation {uuid}")
 
     def ask_stream(self, prompt: str):
-        if 'accessToken' not in ChatGPT.session:
-            self.refresh_session()
+        if 'accessToken' not in self.browser.session:
+            self.browser.refresh_session()
 
         new_message_id = str(uuid.uuid4())
 
-        if "accessToken" not in ChatGPT.session:
+        if "accessToken" not in self.browser.session:
             yield (
                 "Your ChatGPT session is not usable.\n"
                 "* Run this program with the `install` parameter and log in to ChatGPT.\n"
@@ -361,21 +250,21 @@ class ChatGPT:
             };
             xhr.send(JSON.stringify(REQUEST_JSON));
             """.replace(
-                "BEARER_TOKEN", ChatGPT.session["accessToken"]
+                "BEARER_TOKEN", self.browser.session["accessToken"]
             )
             .replace("REQUEST_JSON", json.dumps(request))
             .replace("STREAM_DIV_ID", self.stream_div_id)
             .replace("EOF_DIV_ID", self.eof_div_id)
         )
 
-        self.page.evaluate(code)
+        self.browser.page.evaluate(code)
 
         last_event_msg = ""
         start_time = time.time()
         while True:
-            eof_datas = self.page.query_selector_all(f"div#{self.eof_div_id}")
+            eof_datas = self.browser.page.query_selector_all(f"div#{self.eof_div_id}")
 
-            conversation_datas = self.page.query_selector_all(
+            conversation_datas = self.browser.page.query_selector_all(
                 f"div#{self.stream_div_id}"
             )
             if len(conversation_datas) == 0:
